renamed existing views and reorganized structure of tutorials.
