+++ src/EPiServer.SocialAlloy.Web/Social/Blocks/Ratings/LikeButtonBlock.cs
