+++ src/EPiServer.SocialAlloy.Web/Social/Models/Ratings/LikeButtonBlockViewModel.cs
