--- src/EPiServer.SocialAlloy.Web/Social/Tutorials/README.md
@@ -1,8 +1,0 @@
-﻿# Tutorials
-
-The tutorials are intended to be simple implementations of some typical use cases for
-social features. They have been implemented for educational purposes and are not
-necessarily designed or tested for production use.
-
-For a more comprehensive example of Episerver Social features please refer to the code
-one level up from this folder.
