+++ src/EPiServer.SocialAlloy.Web/Views/Social/RatingBlock/Index.cshtml
